+++ luda/cli.py
@@ -1,17 +1,32 @@
 # -*- coding: utf-8 -*-
+"""
+ideas from:
+https://denibertovic.com/posts/handling-permissions-with-docker-volumes/
+
+"""
 import os
 import subprocess
 
 import click
 
-from luda import which, Volume, add_display, parse_tuple
+from .luda import which, Volume, add_display, parse_tuple, expand_abbreviations, generate_dockerfile_extension
+from .config import read_config
+
 
 PathType = click.Path(exists=True, file_okay=True,
                       dir_okay=True, resolve_path=True)
 
 
+class DockerVolumeType(click.ParamType):
+    name = 'volume'
+
+    def convert(self, value, param, ctx):
+        return Volume.fromString(value)
+
+
 def exclusive(ctx_params, exclusive_params, error_message):
     """
+    Enables defining mutually exclusive options.
     https://gist.github.com/thebopshoobop/51c4b6dce31017e797699030e3975dbf
     
     :param ctx_params: 
@@ -21,34 +36,6 @@ def exclusive(ctx_params, exclusive_params, error_message):
     """
     if sum([1 if ctx_params[p] else 0 for p in exclusive_params]) > 1:
         raise click.UsageError(error_message)
-
-# ideas from:
-# https://denibertovic.com/posts/handling-permissions-with-docker-volumes/
-
-# volumes have the same syntax as docker volumes; however,
-# - relative paths can be resolved
-# - if no mount path is given, the basename of the host path is used, eg
-#   /basename(hostpath)
-#
-# Examples:
-# --volume /path/data:/data:ro
-# --volume data:/data:ro
-# --volume data::ro
-
-# create init/entrypoint script to gnerate a user with a uid inside the
-# contaienr
-# 1) create temp dire with templated init script
-# 2) volume mount tmp dir to /bootstrap
-# 3) set --entrypoint to /bootstrap/init.sh
-# #
-# # Package
-
-
-class DockerVolumeType(click.ParamType):
-    name = 'volume'
-
-    def convert(self, value, param, ctx):
-        return Volume.fromString(value)
 
 
 @click.command(context_settings=dict(
@@ -67,6 +54,7 @@ class DockerVolumeType(click.ParamType):
               help="Adds the DEVTOOLS environment variable. Forces 'apt update' and 'apt " +
                    "install -y --no-install-recommends sudo' to be run before launching the " +
                    "container")
+@click.option('--template', multiple=True, help="Apply template to extend the named image")
 @click.option('--rm', is_flag=True, help="Automatically remove the container when it exits (incompatible with -d)")
 @click.option('-t', '--tty', is_flag=True, help="Allocate a pseudo-tty")
 @click.option('-i', '--stdin', is_flag=True, help="Keep STDIN open even if not attached")
@@ -74,11 +62,12 @@ class DockerVolumeType(click.ParamType):
               help="Detached mode: Run container in the background, print new container id")
 @click.argument('docker_args', nargs=-1, type=click.UNPROCESSED)
 def main(docker_args, display, docker, dev, rm=None, detach=None, tty=None, stdin=None,
-         work=None, home=None, volume=None):
+         template=None, work=None, home=None, volume=None):
     """Console script for luda.
 
     For best results, use a `--` before the image name to ensure all arguments after the image are ignored by luda.
     """
+    config = read_config()
     exclusive(click.get_current_context().params, ['detach', 'rm'], 'd and rm are mutually exclusive')
 
     # if no run options are given, set defaults
@@ -106,6 +95,7 @@ def main(docker_args, display, docker, dev, rm=None, detach=None, tty=None, stdi
     exe = which("nvidia-docker") or "docker"
     args = ["run"]
 
+    # run arguments
     if rm:
         args.append("--rm")
     if detach:
@@ -115,7 +105,7 @@ def main(docker_args, display, docker, dev, rm=None, detach=None, tty=None, stdi
     if stdin:
         args.append("-i")
 
-
+    # override the entrypoint with luda's custom bootstrap
     entrypoint_str = " --entrypoint /bootstrap/init.sh" \
                      " --env HOST_USER_ID={uid}" \
                      " --env HOST_GROUP_ID={gid}" \
@@ -124,11 +114,13 @@ def main(docker_args, display, docker, dev, rm=None, detach=None, tty=None, stdi
     args.append(bootstrap_vol.string)
     args.append(entrypoint_str)
 
+    # map in the user's home directory [optional]
     if home:
         home_vol = Volume("~", "/home/{0}".format(user))
         if home_vol.host_path not in [v.host_path for v in volume]:
             args.append(home_vol.string)
 
+    # map in the current working directory - can be overridden or ignored
     work_vol = None
     if work is None:
         work_vol = Volume(os.getcwd(), "/work")
@@ -140,11 +132,7 @@ def main(docker_args, display, docker, dev, rm=None, detach=None, tty=None, stdi
     if display:
         args.append(add_display())
 
-    if dev:
-        args.append("--env DEVTOOLS=1")
-
-    nvargs = [exe] + [v.string for v in volume] + args
-
+    # look at the remaining arugments to find the container image name
     image_and_args = ()
     docker_args_iter = iter(docker_args)
 
@@ -158,20 +146,19 @@ def main(docker_args, display, docker, dev, rm=None, detach=None, tty=None, stdi
         # otherwise, add it to the image and image args
         image_and_args += (d_arg,)
 
-    # Get the docker command if no args were specified
-    if len(image_and_args) == 1:
-        ep_str = subprocess.Popen([exe, 'inspect', '-f "{{.Config.Entrypoint}}"', image_and_args[0]],
+    # expand image_name if abbreviations are present
+    abbreviations = config.get("abbreviations", {})
+    image_name = expand_abbreviations(image_and_args[0], abbreviations)
+
+    # Determine if the container image has an entrypoint
+    ep_str = subprocess.Popen([exe, 'inspect', '-f "{{.Config.Entrypoint}}"', image_name],
                                   stdout=subprocess.PIPE).stdout.read()
+    ep_str = parse_tuple(ep_str)
 
-        # click.echo(ep_str)
-        ep_str = parse_tuple(ep_str)
-
-        # add the entry point if it exists
-        if ep_str and len(ep_str) > 0:
-            docker_args += (ep_str,)
-
-        # koutputs an array of cmds
-        curr_cmd = subprocess.Popen([exe, 'inspect', '-f "{{.Config.Cmd}}"', image_and_args[0]],
+    # if no default commands are given, inspect the container image for default commands
+    if len(image_and_args) == 1:
+        # outputs an array of cmds
+        curr_cmd = subprocess.Popen([exe, 'inspect', '-f "{{.Config.Cmd}}"', image_name],
                                     stdout=subprocess.PIPE).stdout.read()
         # click.echo(curr_cmd)
         curr_cmd = parse_tuple(curr_cmd)
@@ -179,7 +166,28 @@ def main(docker_args, display, docker, dev, rm=None, detach=None, tty=None, stdi
         if curr_cmd:
             docker_args += (curr_cmd,)
 
-    cmd = " ".join(nvargs + list(docker_args))
+    # update the image_name in the actual argument list in case the image_name was abbreviated
+    docker_args = list(docker_args)
+    image_index = next((idx for idx, arg in enumerate(docker_args) if arg == image_and_args[0]), -1)
+    docker_args[image_index] = image_name
+
+    # if the container image contained an entrypoint, add it immediately after the image name
+    if ep_str and len(ep_str) > 0:
+        docker_args.insert(image_index+1, ep_str)
+
+    # generate dev template and adjust the image_name
+    if dev:
+        image = generate_dockerfile_extension(docker_args[image_index], "dev")
+        docker_args[image_index] = image
+
+    # generate templates in order they are entered on the commandline
+    for t in template:
+        image = generate_dockerfile_extension(docker_args[image_index], t)
+        docker_args[image_index] = image
+
+    # generate docker commandline and execute it (this should probably be an exec instead of a subprocess)
+    nvargs = [exe] + [v.string for v in volume] + args + docker_args
+    cmd = " ".join(nvargs)
     click.echo(cmd)
     subprocess.call(cmd, shell=True)